Create midi-gig-recorder
Initial version coded by gemini with the following prompt : 
can you write a python script that uses the rtmidi module, and that records all MIDI trafic on 2 interfaces, with a timecode.

On the first  interface, standard MIDI commands are received (Start, Stop, PC, CC, Notes...)? We can also receive Sysex messages.

In the record, commands must appear in an intelligible format. For example
<timecode> Start
<timecode> PC <pc_value>
<timecode> CC <cc_number> <value>
<timecode> Note On C3 <velocity>
<timecode> Sysex <raw data in hexadecimal aligned on 2 characeters per byte>

On the 2nd interface, messages are sent by the Allen & Heat CQ18T mixer which use a specific protocol. These CQ18T messages are 9 or 12 bytes messages.

- 9 bytes messages are built as follows: 
   - "mute toggle": 0xB0 0x63 <channel_msb> 0xB0 0x62 <channel_lsb> 0xB0 0x60 0x00

- 12 bytes messages are built as follows:
   - "mute On": 0xB0 0x63 <channel_msb> 0xB0 0x62 <channel_lsb> 0xB0 0x06 0x00 0xB0 0x26 0x01
   - "mute Off": 0xB0 0x63 <channel_msb> 0xB0 0x62 <channel_lsb> 0xB0 0x06 0x00 0xB0 0x26 0x00
   - "level": 0xB0 0x63 <channel_msb> 0xB0 0x62 <channel_lsb> 0xB0 0x06 <value_msb> 0xB0 0x26 <value_lsb>